requirements.txt
baidu-aip==2.2.18.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Pillow==7.2.0
requests==2.24.0
selenium==3.141.0
urllib3==1.25.10
